chore: BankState를 InputOutputManager에서 분리
