+++ BankManagerConsoleApp/BankManagerConsoleApp/BankSystem/InputOutputManager/BankState.swift
@@ -1,5 +1,5 @@
 //
-//  InputOutputManager.swift
+//  BankState.swift
 //  BankManagerConsoleApp
 //
 //  Created by DONGWOOK SEO on 2023/02/27.
@@ -7,26 +7,13 @@
 
 import Foundation
 
-enum InputOutputManager {
-    static func output(state: BankState) {
-        print(state.text, terminator: "")
-    }
-    
-    static func input() -> String {
-        guard let userInput = readLine() else {
-            return ""
-        }
-        return userInput
-    }
-}
-
 enum BankState {
     case open
     case close(UInt, Double)
     case working(UInt, Bool)
 }
 
-fileprivate extension BankState {
+extension BankState {
     var text: String {
         switch self {
         case .open:
